requirements.txt
discord.py
flask
aiohttp
wikipedia
requests
pytz
datetime
praw
pymongo
dnspython
asyncio
